Handle rendering multiple locations
